Extracted featured query to its own module.
